requirements.txt
crewai
python-dotenv
GitPython
requests